Update auto_stops.py
to ensure the periodic check will not run more than once, introduced a token mechanism for the periodic check to determine if it can continue or not, hence, in effect, no more than one periodic check can coexist behind the scenes. besides, added a matching in `on_modified()` to minimize the "`stops` being bloated up with unnecessary regions due to certain typing scenarios", plus a minor touch of other thing.