Update auto_stops.py
i suddenly realized that if `on_modified` and `check_idle` are running in parallel, the uncertainty due to the complication of such scenario is problematic. so, i decided to prevent such scenario from happening. hence, i added the `threading.Lock()` to regulate them, so that they can't be running at the same time. besides, i also added a cleanup point for the `last_activity` list at `on_pre_close`; as well as added a short-circuit at `on_activated` for all irrelevant views, such as, all sort of panels.